feat: Push update of latestMilestoneIndex and timestamp from Landing down to Feed subclass
