+++ client/src/app/components/stardust/Feeds.tsx
@@ -107,9 +107,25 @@ abstract class Feeds<P extends RouteComponentProps<{ network: string }>, S exten
 
     /**
      * The items have been updated.
-     * @param items The updated items.
+     * @param newItems The updated items.
      */
-    protected itemsUpdated(items: IFeedItem[]): void {
+    protected itemsUpdated(newItems: IFeedItem[]): void {
+        if (newItems) {
+            const milestones = newItems.filter(i => i.payloadType === "MS");
+            for (const ms of milestones) {
+                const index: number | undefined = ms.properties?.index as number;
+                const currentIndex = this.state.latestMilestoneIndex;
+                if (index && currentIndex !== undefined && index > currentIndex) {
+                    const timestamp: number | undefined = ms.properties?.timestamp as number;
+                    if (timestamp) {
+                        this.setState({
+                            latestMilestoneIndex: index,
+                            latestMilestoneTimestamp: timestamp * 1000
+                        });
+                    }
+                }
+            }
+        }
     }
 
     /**
